Move checks tests to new folder
